+++ packages/@magic-sdk/provider/test/constants.ts
@@ -2,7 +2,6 @@ export const MAGIC_RELAYER_FULL_URL = 'https://auth.magic.link';
 export const TEST_API_KEY = 'pk_test_123';
 export const LIVE_API_KEY = 'pk_live_123';
 export const ENCODED_QUERY_PARAMS = 'testqueryparams';
-export const TEST_NETWORK_HASH = 'eth_mainnet';
 export const MSG_TYPES = (parameters = ENCODED_QUERY_PARAMS) => ({
   MAGIC_HANDLE_RESPONSE: `MAGIC_HANDLE_RESPONSE-${parameters}`,
   MAGIC_OVERLAY_READY: `MAGIC_OVERLAY_READY-${parameters}`,
