+++ packages/@magic-sdk/provider/test/spec/core/view-controller/post.spec.ts
@@ -3,7 +3,6 @@ import { createViewController, TestViewController } from '../../../factories';
 import { JsonRpcResponse } from '../../../../src/core/json-rpc';
 import * as storage from '../../../../src/util/storage';
 import * as webCryptoUtils from '../../../../src/util/web-crypto';
-import * as deviceShareWebCryptoUtils from '../../../../src/util/device-share-web-crypto';
 import { SDKEnvironment } from '../../../../src/core/sdk-environment';
 import { createModalNotReadyError } from '../../../../src/core/sdk-exceptions';
 
@@ -22,7 +21,7 @@ function requestPayload(id = 1): JsonRpcRequestPayload {
 /**
  * Create a dummy response payload.
  */
-function responseEvent(values: { result?: any; error?: any; id?: number; deviceShare?: string } = {}) {
+function responseEvent(values: { result?: any; error?: any; id?: number } = {}) {
   return {
     data: {
       response: {
@@ -31,7 +30,6 @@ function responseEvent(values: { result?: any; error?: any; id?: number; deviceS
         jsonrpc: '2.0',
         id: values.id ?? 1,
       },
-      deviceShare: values.deviceShare ?? null,
     },
   };
 }
@@ -62,10 +60,7 @@ function stubViewController(viewController: any, events: [MagicIncomingWindowMes
 }
 
 let createJwtStub: jest.SpyInstance<Promise<string | undefined>>;
-let getDecryptedDeviceShareStub: jest.SpyInstance<Promise<string | undefined>>;;
-let clearDeviceSharesStub: jest.SpyInstance<Promise<void>>;
 const FAKE_JWT_TOKEN = 'hot tokens';
-const FAKE_DEVICE_SHARE = 'fake device share';
 const FAKE_RT = 'will freshen';
 const FAKE_INJECTED_JWT = 'fake injected jwt';
 let FAKE_STORE: Record<string, string> = {};
@@ -76,8 +71,6 @@ beforeEach(() => {
   jest.resetAllMocks();
   jest.restoreAllMocks();
   createJwtStub = jest.spyOn(webCryptoUtils, 'createJwt');
-  getDecryptedDeviceShareStub = jest.spyOn(deviceShareWebCryptoUtils, 'getDecryptedDeviceShare');
-  clearDeviceSharesStub = jest.spyOn(deviceShareWebCryptoUtils, 'clearDeviceShares');
   jest.spyOn(global.console, 'info').mockImplementation(() => { /* noop */ });
   jest.spyOn(global, 'addEventListener').mockImplementation(jest.fn());
   jest.spyOn(storage, 'getItem').mockImplementation((key: string, callback?: (err: unknown, value: unknown) => void) => {
@@ -130,45 +123,6 @@ test('Sends payload with jwt when web crypto is supported', async () => {
   expect(response).toEqual(new JsonRpcResponse(responseEvent().data.response));
   expect(createJwtStub).toHaveBeenCalledWith();
   expect(postSpy).toBeCalledWith(expect.objectContaining({ jwt: FAKE_JWT_TOKEN }));
-});
-
-test('Sends payload with deviceShare when it is saved', async () => {
-  createJwtStub.mockImplementationOnce(() => Promise.resolve(FAKE_JWT_TOKEN));
-  getDecryptedDeviceShareStub.mockImplementationOnce(() => Promise.resolve(FAKE_DEVICE_SHARE));
-  const eventWithDeviceShare = { data: { ...responseEvent().data, deviceShare: FAKE_DEVICE_SHARE } };
-
-  const { postSpy } = stubViewController(viewController, [
-    [MagicIncomingWindowMessage.MAGIC_HANDLE_RESPONSE, eventWithDeviceShare],
-  ]);
-  const payload = requestPayload();
-  await viewController.post(MagicOutgoingWindowMessage.MAGIC_HANDLE_REQUEST, payload);
-
-  expect(createJwtStub).toHaveBeenCalledWith();
-  expect(postSpy).toHaveBeenCalledWith(expect.objectContaining({ deviceShare: FAKE_DEVICE_SHARE }));
-});
-
-test('device share should be cleared if replied user denied account access.', async () => {
-  createJwtStub.mockImplementationOnce(() => Promise.resolve(FAKE_JWT_TOKEN));
-  getDecryptedDeviceShareStub.mockImplementationOnce(() => Promise.resolve(FAKE_DEVICE_SHARE));
-  clearDeviceSharesStub.mockImplementationOnce(() => Promise.resolve());
-  const eventWithError = {
-    data: {
-      ...responseEvent().data,
-      response: {
-        id: 1,
-        error: { message: 'User denied account access.' },
-      },
-    },
-  };
-
-  const { postSpy } = stubViewController(viewController, [
-    [MagicIncomingWindowMessage.MAGIC_HANDLE_RESPONSE, eventWithError],
-  ]);
-  const payload = requestPayload();
-  await viewController.post(MagicOutgoingWindowMessage.MAGIC_HANDLE_REQUEST, payload);
-
-  expect(createJwtStub).toHaveBeenCalledWith();
-  expect(postSpy).toHaveBeenCalledWith(expect.objectContaining({ deviceShare: FAKE_DEVICE_SHARE }));
 });
 
 test('Sends payload with rt and jwt when rt is saved', async () => {
