+++ packages/@magic-sdk/types/src/core/message-types.ts
@@ -28,7 +28,6 @@ export interface MagicMessageRequest {
   payload: JsonRpcRequestPayload | JsonRpcRequestPayload[];
   rt?: string;
   jwt?: string;
-  deviceShare?: string;
 }
 
 /** The shape of responding window message data from the Magic iframe context. */
@@ -36,7 +35,6 @@ export interface MagicMessageResponse<ResultType = any> {
   msgType: string;
   response: Partial<JsonRpcError> & Partial<JsonRpcResponsePayload<ResultType>>;
   rt?: string;
-  deviceShare?: string;
 }
 
 /** The expected message event returned by the Magic iframe context. */
